Modified sim_null for updating the library of lambda each time a prediction is made using METE. Modified create_null_dataset to incorporate the library of lambda.
Kate - Now to run this you probably need to create an empty library in the wrapper the first time you run this with the following lines: 

import pickle

lib_lambda = {}
file = open('lib_name.pck', 'w')
pickle.dump(lib_lambda, file)
file.close()

This only needs to be done for the first dataset (or outside the whole analysis, if you prefer). The library will get updated simulatenously with the iterations and hopefully can be directly read in for later analyses.