Add missing parentheses to properly subset FIA data
This fixes a bug in one of the FIA queries that results in a
slightly incorrect subset of the FIA data.

From Kate Thibault:
"The results include violations of two of the conditions:
QA_STATUS and SAMP_METHOD_CD. The QA_STATUS issue is that hot
checks of the plots are included (status = 7) – recounts of the
plots in the field to check on the data. The SAMP_METHOD_CD issue
is that there are some with cd = 2 – remotely sensed plot."

Analyses have been rerun and are indistinguishable from the
reported results.